+++ lib/national_assembly_france_hemicycle.dart
@@ -1,5 +1,8 @@
-export 'attic/json_vote_objecter.dart';
-export 'attic/json_transcoder.dart';
-export 'attic/colors.dart';
-export 'attic/helpers.dart';
 export 'attic/a_n_files_fetch.dart';
+export 'attic/colors.dart';
+export 'attic/folders.dart';
+export 'attic/groupe_transcode.dart';
+export 'attic/helpers.dart';
+export 'attic/json_transcoder.dart';
+export 'attic/json_vote_objecter.dart';
+export 'attic/votes_displayer.dart';